[ Add ]：Add string extension and label extension to SwiftMinions.
